src-backend/requirements.txt
pyspark==3.5.5
narwhals
hdfs
python-dotenv